requirements.txt
requests>=2.7.0
six>=1.8.0
furl>=0.4.6
optionaldict>=0.1.0
python-dateutil>=2.4.2
